Add integration tests for likes and comments
