+++ spec/examples.txt
@@ -1,20 +1,23 @@
 example_id                                       | status | run_time        |
 ------------------------------------------------ | ------ | --------------- |
-./spec/comments_unit_test_spec.rb[1:1:1]         | passed | 0.00147 seconds |
-./spec/comments_unit_test_spec.rb[1:1:2]         | passed | 0.00627 seconds |
-./spec/comments_unit_test_spec.rb[1:1:3]         | passed | 0.00743 seconds |
-./spec/comments_unit_test_spec.rb[1:1:4]         | passed | 0.00931 seconds |
-./spec/likes_unit_test_spec.rb[1:1:1]            | passed | 0.00743 seconds |
-./spec/likes_unit_test_spec.rb[1:1:2]            | passed | 0.04257 seconds |
-./spec/likes_unit_test_spec.rb[1:1:3]            | passed | 0.02174 seconds |
-./spec/post_unit_test_spec.rb[1:1:1]             | passed | 0.00175 seconds |
-./spec/post_unit_test_spec.rb[1:1:2]             | passed | 0.00518 seconds |
-./spec/post_unit_test_spec.rb[1:1:3]             | passed | 0.00502 seconds |
-./spec/user_signin_integration_test_spec.rb[1:1] | passed | 0.07139 seconds |
-./spec/user_signin_integration_test_spec.rb[1:2] | passed | 0.0494 seconds  |
-./spec/user_signup_integration_test_spec.rb[1:1] | passed | 0.07115 seconds |
-./spec/user_signup_integration_test_spec.rb[1:2] | passed | 0.31163 seconds |
-./spec/user_unit_test_spec.rb[1:1:1]             | passed | 0.00315 seconds |
-./spec/user_unit_test_spec.rb[1:1:2]             | passed | 0.00276 seconds |
-./spec/user_unit_test_spec.rb[1:1:3]             | passed | 0.00341 seconds |
-./spec/user_unit_test_spec.rb[1:1:4]             | passed | 0.00236 seconds |
+./spec/comments_integration_test_spec.rb[1:1]    | passed | 0.06087 seconds |
+./spec/comments_unit_test_spec.rb[1:1:1]         | passed | 0.00177 seconds |
+./spec/comments_unit_test_spec.rb[1:1:2]         | passed | 0.00606 seconds |
+./spec/comments_unit_test_spec.rb[1:1:3]         | passed | 0.00676 seconds |
+./spec/comments_unit_test_spec.rb[1:1:4]         | passed | 0.03733 seconds |
+./spec/likes_integration_test_spec.rb[1:1]       | passed | 0.05902 seconds |
+./spec/likes_integration_test_spec.rb[1:2]       | passed | 0.06225 seconds |
+./spec/likes_unit_test_spec.rb[1:1:1]            | passed | 0.01418 seconds |
+./spec/likes_unit_test_spec.rb[1:1:2]            | passed | 0.00824 seconds |
+./spec/likes_unit_test_spec.rb[1:1:3]            | passed | 0.00735 seconds |
+./spec/post_unit_test_spec.rb[1:1:1]             | passed | 0.00131 seconds |
+./spec/post_unit_test_spec.rb[1:1:2]             | passed | 0.00523 seconds |
+./spec/post_unit_test_spec.rb[1:1:3]             | passed | 0.00477 seconds |
+./spec/user_signin_integration_test_spec.rb[1:1] | passed | 0.04681 seconds |
+./spec/user_signin_integration_test_spec.rb[1:2] | passed | 0.06044 seconds |
+./spec/user_signup_integration_test_spec.rb[1:1] | passed | 0.28088 seconds |
+./spec/user_signup_integration_test_spec.rb[1:2] | passed | 0.03938 seconds |
+./spec/user_unit_test_spec.rb[1:1:1]             | passed | 0.00281 seconds |
+./spec/user_unit_test_spec.rb[1:1:2]             | passed | 0.00254 seconds |
+./spec/user_unit_test_spec.rb[1:1:3]             | passed | 0.00282 seconds |
+./spec/user_unit_test_spec.rb[1:1:4]             | passed | 0.00245 seconds |
