+++ spec/user_signin_integration_test_spec.rb
@@ -20,15 +20,4 @@ RSpec.describe 'User signin', type: :feature do
     click_button 'Log in'
     expect(page).to have_content('Log in')
   end
-
-  def test_user_signup
-    visit new_user_registration_path
-    fill_in 'Name', with: 'Name'
-    fill_in 'Birth date', with: DateTime.now
-    page.choose('user_gender_male')
-    fill_in 'Email', with: 'mail@mail.com'
-    fill_in 'Password', with: 'Pass123'
-    fill_in 'Password confirmation', with: 'Pass123'
-    click_button 'Sign up'
-  end
 end
